Calculate coverage by chunk, aggregate per genome
