+++ travisci/perl-unittest_harness.sh
@@ -40,7 +40,7 @@ find docs modules scripts sql travisci -iname '*.t' -print0 | xargs -0 -n 1 perl
 rt2=$?
 find docs modules scripts sql travisci -iname '*.pl' -print0 | xargs -0 -n 1 perl -c
 rt3=$?
-find docs modules scripts sql travisci -iname '*.pm' \! -name 'LoadSynonyms.pm' \! -name 'HALAdaptor.pm' \! -name 'HALXS.pm' \! -name 'halSequenceFactory.pm' -print0 | xargs -0 -n 1 perl -c
+find docs modules scripts sql travisci -iname '*.pm' \! -name 'LoadSynonyms.pm' \! -name 'HALAdaptor.pm' \! -name 'HALXS.pm' -print0 | xargs -0 -n 1 perl -c
 rt4=$?
 
 if [[ ($rt1 -eq 0) && ($rt2 -eq 0) && ($rt3 -eq 0) && ($rt4 -eq 0)]]; then
